flash before deleting a user if deputy

+++ src/Controller/UserController.php
@@ -127,6 +127,12 @@ class UserController extends AbstractController
 
             return $this->redirectToRoute('user_index');
         }
+//        dd($user);
+        if($user->getAssociatedWith() !== null ) {
+            $this->addFlash("error", "Veuillez retirer le suppléant avant de supprimer cet utilisateur");
+            return $this->redirectToRoute('user_index');
+
+        }
         $manageUser->delete($user);
         $this->addFlash('success', 'L\'utilisateur a bien été supprimé');
